add product to checkout
