+++ src/components/Checkout/ContactInformation/ContactInformation.module.css
@@ -1,0 +1,90 @@
+.title {
+  @apply px-6
+  py-3
+  border-b
+  border-b-border
+  text-text_black
+  text-title_checkout
+  font-title_checkout
+  leading-title_checkout;
+}
+
+.titleSecond {
+  @apply mt-6
+  
+  text-text_black
+  text-title_second
+  font-title_second
+  leading-title_second;
+}
+
+.inputText {
+  @apply px-1 py-2 shadow-lg;
+}
+
+.wrapper {
+  @apply bg-blue-500 p-4 rounded-lg shadow-md;
+}
+
+.text {
+  @apply text-white text-center mt-2;
+}
+
+.label {
+  @apply ml-2 block text-sm font-medium leading-6 text-gray-500;
+}
+
+.input {
+  @apply block px-2 py-1.5 w-full rounded border-0  text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 outline-none;
+}
+
+.select {
+  @apply block px-2 py-1.5 w-full rounded border-0  text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 outline-none;
+}
+
+.deliveryWrapper {
+  @apply mt-4 relative inline-block text-left;
+}
+
+.wrapButton {
+  @apply flex justify-between
+  mt-6
+  pb-2
+  text-type_add_info
+  font-type_add_info
+  leading-type_add_info;
+}
+
+.wrapButton .buttonCheckout {
+  @apply px-8
+  py-3
+  min-w-[264px]
+  rounded-lg 
+  bg-button_dark_background
+  text-white
+  
+  hover:text-action_color
+   hover:shadow-gray-500/80
+  hover:shadow-lg  
+
+  /* focus-visible:outline 
+  focus-visible:outline-2 
+  focus-visible:outline-offset-2 
+  focus-visible:outline-indigo-600  */
+  transition;
+}
+
+.buttonBack {
+  @apply flex gap-3 
+  py-3
+  
+  hover:text-action_color;
+}
+
+.iconBack {
+  @apply text-icon_back_color;
+}
+
+.buttonBack:hover .iconBack {
+  @apply text-action_color;
+}
