+++ src/screens/Checkout/Checkout.module.css
@@ -1,27 +1,11 @@
-.inputText {
-  @apply px-1 py-2 shadow-lg;
+.wrapComponent {
+  @apply flex gap-6;
 }
 
-.wrapper {
-  @apply bg-blue-500 p-4 rounded-lg shadow-md;
+.wrapContactInformation {
+  @apply grow;
 }
 
-.text {
-  @apply text-white text-center mt-2;
-}
-
-.label {
-  @apply ml-2 block text-sm font-medium leading-6 text-gray-500;
-}
-
-.input {
-  @apply block px-2 py-1.5 w-full rounded border-0  text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 outline-none;
-}
-
-.select {
-  @apply block px-2 py-1.5 w-full rounded border-0  text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6 outline-none;
-}
-
-.deliveryWrapper {
-  @apply mt-4 relative inline-block text-left;
+.wrapProducts {
+  @apply pt-14 w-[503px];
 }
